Remove files that shouldn't be in the moira tree: these should already be installed, and we should use the installed versions.
Update remaining include files for 64-bit OSes.

+++ moira/include/mr_proto.h
@@ -1,7 +1,7 @@
 /*
  *	$Source: /afs/.athena.mit.edu/astaff/project/moiradev/repository/moira/include/mr_proto.h,v $
- *	$Author: mar $
- *	$Header: /afs/.athena.mit.edu/astaff/project/moiradev/repository/moira/include/mr_proto.h,v 1.5 1990-03-17 16:31:32 mar Exp $
+ *	$Author: danw $
+ *	$Header: /afs/.athena.mit.edu/astaff/project/moiradev/repository/moira/include/mr_proto.h,v 1.6 1997-01-29 23:21:23 danw Exp $
  *
  *	Copyright (C) 1987, 1990 by the Massachusetts Institute of Technology
  *
@@ -12,8 +12,8 @@
 #include <moira.h>
 
 typedef struct mr_params {
-	u_long mr_size;
-	u_long mr_version_no;
+	uint32 mr_size;
+	uint32 mr_version_no;
 	union {
 		u_long procno;	/* for call */
 		u_long status;	/* for reply */
